DIfferent enemy & poweup added

+++ Assets/Scripts/SceneEvents/OnPlayerTankKilled/OnPlayerTankKilled.cs
@@ -6,7 +6,7 @@ namespace SceneEvents
     public class OnPlayerTankKilled : SceneEvent
     {
         [SerializeField] private Spawner _spawner;
-        [SerializeField] private int _playerSpawnTime = 3000;
+        [SerializeField] private float _playerSpawnTime = 3000;
 
 
         public override void TriggerEvent()
